Switched CMakeLists from ${OPENGL_gl_LIBRARY} to CMake built-in OpenGL::GL definition.
Removed some unused messsages.
Changed prefixScreenshot system option default location to the current working directory.
(Avoids DevIL errors when the destination folder was not existing.)

+++ apps/rtigo3/dp/math/src/Math.cpp
@@ -94,7 +94,7 @@ namespace dp
         {
           exponent = - exponent;
         }
-        f *= pow( 10.0f, float(exponent) );
+        f *= powf( 10.0f, float(exponent) );
       }
     #if !defined( NDEBUG )
       float z = (float)atof( str.c_str() );
